Integrate mkdocs-material search framework with sphinx's search index
The changes are as follows:

mkdocs uses lunr for searching, which requires loading a rather large
search index json file, and if the "prebuilt index" mode is used, it
can be significantly larger even than the full text content of the
website.  This change replaces the use of lunr with a new search
implementation derived from the searchtools.js file in upstream
sphinx; this new search implementation uses the unmodified
searchindex.js produced by Sphinx.

mkdocs-material shows results for individual sections of a page, but
just shows the start of the section as a snippet.  The section title
and content (to produce a snippet) are all available directly in the
lunr index.  Sphinx uses a significantly smaller index, because it
does not contain the full text itself. In order to show snippets after
finding the results, the upstream searchtools.js (and also this custom
implementation based on it) fetches the actual target html document
for each result, extracts the main content and finds the first
occurrence of a search term. If it fails to fetch the
document (e.g. due to being run from a file:// URL) then it just
doesn't show a snippet.  This new search implementation also does
that, but additionally parses the sections in order to include section
titles in the results and provide result links directly to individual
sections.

The sphinx search index includes a special index for "objects",
e.g. classes, methods, etc.  These are supported as well, but no
snippet extraction is done for those results.

Since extracting the section titles and snippets requires retrieving
the actual target documents, for a large number of results this could
take a while. To mitigate that, fetching is controlled by the scroll
position of the results view: snippets are only fetched for visible
results, and then more are fetched if the user scrolls down.