Integrate mkdocs-material's node.js-based build system into sphinx-material
mkdocs-material uses a node.js-based build system to build the
javascript and css bundles, to include the appropriate bundle paths in
the HTML jinja templates, and to copy icons.

This commit modifies the mkdocs-material build system to work with
sphinx-material.

mkdocs-material copies several entire svg icon repositories into its
package, which amounts to about 30MiB.  To avoid bloating the
published package size for all users, this change modifies the build
system to determine the set of icons actually referenced by SCSS or
HTML templates, and only copies those.

Additionally, this commit adds logic to setup.py to automatically
invoke the node.js build system as part of the build_py command, which
is run automatically when creating an sdist, building a wheel, or
installing.